refactor(product-picker): shared -> intenals
