+++ packages/application-components/src/experimental/components/internals/dropdown-indicator/dropdown-indicator.tsx
