+++ packages/application-components/src/experimental/components/internals/dropdown-indicator/index.ts
